Move utils/logging -> logging
