+++ mlprogram/actions/action_sequence.py
@@ -6,7 +6,7 @@ from mlprogram.actions.action \
     import Action, ApplyRule, Rule, ExpandTreeRule, NodeConstraint, NodeType, \
     GenerateToken, CloseVariadicFieldRule
 from mlprogram.asts import AST, Node, Leaf, Field, Root
-from mlprogram.utils import logging
+from mlprogram import logging
 
 
 logger = logging.Logger(__name__)
