+++ mlprogram/actions/ast_to_action_sequence.py
@@ -5,7 +5,7 @@ from mlprogram.actions import ApplyRule, ExpandTreeRule
 from mlprogram.actions import CloseVariadicFieldRule
 from mlprogram.actions import GenerateToken
 from mlprogram.actions import ActionSequence
-from mlprogram.utils import logging
+from mlprogram import logging
 
 
 logger = logging.Logger(__name__)
