+++ mlprogram/datasets/hearthstone/download.py
@@ -1,7 +1,7 @@
 import requests
 from typing import Callable, Dict
 
-from mlprogram.utils import logging
+from mlprogram import logging
 from mlprogram.utils.data import ListDataset
 
 logger = logging.Logger(__name__)
