+++ mlprogram/languages/csg/functions.py
@@ -8,7 +8,7 @@ from mlprogram.encoders import Samples
 from mlprogram.actions \
     import ActionSequence, ApplyRule, CloseVariadicFieldRule, Rule
 from mlprogram.asts import AST, Node, Field, Leaf, Root
-from mlprogram.utils import logging
+from mlprogram import logging
 
 logger = logging.Logger(__name__)
