+++ mlprogram/samplers/action_sequence_sampler.py
@@ -15,7 +15,7 @@ from mlprogram.samplers import SamplerState, DuplicatedSamplerState, Sampler
 from mlprogram.nn.utils.rnn import PaddedSequenceWithMask
 from mlprogram.utils import TopKElement, Token
 from mlprogram.utils.data import Collate
-from mlprogram.utils import logging
+from mlprogram import logging
 
 logger = logging.Logger(__name__)
