+++ mlprogram/samplers/ast_reference_sampler.py
@@ -12,7 +12,7 @@ from mlprogram.interpreters import SequentialProgram
 from mlprogram.synthesizers import Synthesizer
 from mlprogram.utils import Token
 from mlprogram.utils.data import Collate
-from mlprogram.utils import logging
+from mlprogram import logging
 
 logger = logging.Logger(__name__)
