+++ mlprogram/samplers/sampler_with_value_network.py
@@ -3,7 +3,7 @@ from typing \
     import TypeVar, Generic, Generator, Optional, List, Callable, Any, Tuple
 from mlprogram.samplers import SamplerState, Sampler, DuplicatedSamplerState
 from mlprogram.utils.data import Collate
-from mlprogram.utils import logging
+from mlprogram import logging
 
 logger = logging.Logger(__name__)
