+++ mlprogram/synthesizers/beam_search.py
@@ -1,7 +1,7 @@
 from typing import TypeVar, Generic, Generator, Optional
 from mlprogram.samplers import Sampler, SamplerState
 from mlprogram.synthesizers import Synthesizer, Result
-from mlprogram.utils import logging
+from mlprogram import logging
 
 logger = logging.Logger(__name__)
