+++ mlprogram/synthesizers/filtered_synthesizer.py
@@ -1,6 +1,6 @@
 from typing import TypeVar, Generic, Generator, Callable, Optional
 from mlprogram.synthesizers import Synthesizer, Result
-from mlprogram.utils import logging
+from mlprogram import logging
 
 logger = logging.Logger(__name__)
